move download tests from starter to utils
